requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt
django-filter
django-environ==0.11.2
django-cors-headers
python-jose
pytest-django==4.8.0
pytest-mock==3.12.0
pytest-factoryboy
pytest
pillow